+++ scripts/build.sh
@@ -6,4 +6,4 @@ if [ ! -f $CONFIG_FILE ]
     CONFIG_FILE='node_modules/cultureamp-front-end-scripts/config/webpack/webpack.config.js'
 fi
 
-node_modules/.bin/webpack --mode production --config $CONFIG_FILE
+NODE_ENV=production node_modules/.bin/webpack --mode production --config $CONFIG_FILE